Reorganize template-related code - consolidate in models.bots.templates - put AST stuff in models.bots.templates.ast
