+++ app/models/bots/Behavior.scala
@@ -3,10 +3,9 @@ package models.bots
 import java.nio.ByteBuffer
 import java.nio.charset.Charset
 import java.util
-
-import _root_.util.TemplateApplier
 import com.github.tototoshi.slick.PostgresJodaSupport._
 import models.accounts.User
+import models.bots.templates.{SlackRenderer, TemplateApplier}
 import models.{EnvironmentVariableQueries, IDs, Team}
 import org.commonmark.ext.autolink.AutolinkExtension
 import org.commonmark.ext.gfm.strikethrough.StrikethroughExtension
@@ -15,7 +14,6 @@ import org.commonmark.parser.Parser
 import org.joda.time.DateTime
 import play.api.libs.json.{JsString, JsDefined, Json, JsValue}
 import play.api.{Configuration, Play}
-import renderers.SlackRenderer
 import services.AWSLambdaConstants._
 import services.AWSLambdaService
 import slick.driver.PostgresDriver.api._
