+++ app/models/bots/templates/ast/Expr.scala
@@ -1,9 +1,7 @@
-package utils
+package models.bots.templates.ast
 
+import models.bots.templates.MarkdownRenderer
 import play.api.libs.json.JsLookupResult
-import renderers.MarkdownRenderer
-
-import scala.util.parsing.combinator.{JavaTokenParsers, RegexParsers}
 
 sealed trait Expr {
 
@@ -70,31 +68,5 @@ case class Iteration(item: Identifier, list: Path, body: Block) extends Expr {
   override def accept(visitor: MarkdownRenderer): Unit = {
     visitor.visit(this)
   }
-
-}
-
-class TemplateParser extends RegexParsers with JavaTokenParsers {
-
-  override def skipWhitespace = false
-
-  val reserved: Parser[String] = "endfor"
-
-  def text: Parser[Text] = """[^\{]+""".r ^^ { s => Text(s) }
-
-  def identifier: Parser[Identifier] = not(reserved) ~> ident ^^ { s => Identifier(s) }
-
-  def path: Parser[Path] = repsep(identifier, ".") ^^ { segments => Path(segments) }
-
-  def substitution: Parser[Substitution] = """\{\s*""".r ~> path <~ """\s*\}""".r ^^ { case path => Substitution(path) }
-
-  def block: Parser[Block] = rep(text | substitution | iteration) ^^ { children => Block(children) }
-
-  def iteration: Parser[Iteration] =
-    """\{\s*for\s+""".r ~> identifier ~ """\s+in\s+""".r ~ path ~ """\s*\}""".r ~ block <~ """\{\s*endfor\s*\}""".r ^^ {
-    case itemIdentifier ~ _ ~ listPath ~ _ ~ block =>
-      Iteration(itemIdentifier, listPath, block)
-  }
-
-  def parseBlockFrom(text: String): ParseResult[Block] = parse(block, text)
 
 }
